+++ frontend/src/hooks/board/useBoardZoom.ts
