+++ frontend/src/pages/ManagerMapEditor/units/MapEditor.tsx
@@ -8,17 +8,17 @@ import ColorPicker from 'components/ColorPicker/ColorPicker';
 import ColorPickerIcon from 'components/ColorPicker/ColorPickerIcon';
 import { EDITOR, KEY } from 'constants/editor';
 import PALETTE from 'constants/palette';
+import useBindKeyPress from 'hooks/board/useBindKeyPress';
 import useBoardCoordinate from 'hooks/board/useBoardCoordinate';
 import useBoardMove from 'hooks/board/useBoardMove';
 import useBoardStatus from 'hooks/board/useBoardStatus';
+import useBoardZoom from 'hooks/board/useBoardZoom';
 import { Color, DrawingStatus, ManagerSpace, MapElement } from 'types/common';
 import { MapElementType, MapEditorMode } from 'types/editor';
-import useBindKeyPress from '../../../hooks/board/useBindKeyPress';
 import useBoardEraserTool from '../hooks/useBoardEraserTool';
 import useBoardLineTool from '../hooks/useBoardLineTool';
 import useBoardRectTool from '../hooks/useBoardRectTool';
 import useBoardSelect from '../hooks/useBoardSelect';
-import useBoardZoom from '../hooks/useBoardZoom';
 import Board from './Board';
 import * as Styled from './MapEditor.styles';
